% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_precommit_hook_config}
\alias{use_precommit_hook_config}
\title{Copy a default pre-commit-config.yaml in the root of your project}
\usage{
use_precommit_hook_config(hooks = NULL, path_root = here::here())
}
\arguments{
\item{hooks}{Ignored.}

\item{path_root}{The path to the root directory of your project.}
}
\description{
Copy a default pre-commit-config.yaml in the root of your project
}
